fix: PR review changes and paparazzi test
