+++ uicomponents/src/test/java/com/valkiria/uicomponents/bricks/FilledTextFieldViewPaparazziTest.kt
@@ -1,0 +1,24 @@
+package com.valkiria.uicomponents.bricks
+
+import app.cash.paparazzi.DeviceConfig
+import app.cash.paparazzi.Paparazzi
+import com.android.ide.common.rendering.api.SessionParams
+import org.junit.Rule
+import org.junit.Test
+
+class FilledTextFieldViewPaparazziTest {
+
+    @get:Rule
+    val paparazziRule: Paparazzi = Paparazzi(
+        theme = "android:Theme.MaterialComponents.Light.NoActionBar",
+        deviceConfig = DeviceConfig.PIXEL_6_PRO.copy(softButtons = false, screenHeight = 1),
+        renderingMode = SessionParams.RenderingMode.V_SCROLL
+    )
+
+    @Test
+    fun snapFilledTextFieldView() {
+        paparazziRule.snapshot {
+            FilledTextFieldViewPreview()
+        }
+    }
+}
